Move P2LCommunicator to 'dtos' package
